requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
idna==2.6
Naked==0.1.31
pycparser==2.18
pycrypto==2.6.1
PyYAML==3.12
requests==2.18.4
shellescape==3.4.1
six==1.11.0
urllib3==1.22
